% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trekfont.R
\name{show_trekfonts}
\alias{show_trekfonts}
\title{Show the complete font set}
\usage{
show_trekfonts()
}
\value{
a character vector.
}
\description{
List all available Star Trek fonts contained in the package.
}
\examples{
show_trekfonts()
}
