% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trekfont.R
\docType{package}
\name{trekfont}
\alias{trekfont-package}
\alias{trekfont}
\title{trekfont: Star Trek Fonts Collection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a collection of true type and open type Star Trek-themed fonts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leonawicz/trekfont}
  \item Report bugs at \url{https://github.com/leonawicz/trekfont/issues}
}

}
\author{
\strong{Maintainer}: Matthew Leonawicz \email{rpkgs@pm.me} (\href{https://orcid.org/0000-0001-9452-2771}{ORCID})

}
