% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trekfont.R
\docType{data}
\name{trekfonts}
\alias{trekfonts}
\title{Available Star Trek fonts.}
\format{A character vector of 107 font files.}
\usage{
trekfonts
}
\description{
Comprehensive listing of all available fonts. These fonts are sourced from https://github.com/wrstone/fonts-startrek under the GPL-3 license.
}
\keyword{datasets}
